# contracts/openapi/pms-api.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PMS Backend API",
    "version": "1.0.0",
    "description": "Contrat initial publié par pms-backend pour le BFF pms-frontend."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "security": [
    {
      "keycloakBearer": []
    }
  ],
  "paths": {
    "/accounting/opening-balances/certifications": {
      "post": {
        "operationId": "certifyOpeningBalances",
        "parameters": [
          {
            "$ref": "#/components/parameters/CorrelationId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "certificationId",
                  "signedAt",
                  "lines"
                ],
                "properties": {
                  "certificationId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "signedAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "lines": {
                    "type": "array",
                    "minItems": 4,
                    "maxItems": 4,
                    "items": {
                      "type": "object",
                      "required": [
                        "component",
                        "currencyCode",
                        "migratedAmount",
                        "generalLedgerAmount",
                        "evidenceReference"
                      ],
                      "properties": {
                        "component": {
                          "type": "string",
                          "enum": [
                            "HISTORICAL_ACCOUNTS",
                            "PER",
                            "IRR",
                            "PAST_DISTRIBUTIONS"
                          ]
                        },
                        "currencyCode": {
                          "type": "string",
                          "pattern": "^[A-Z]{3}$"
                        },
                        "migratedAmount": {
                          "type": "string"
                        },
                        "generalLedgerAmount": {
                          "type": "string"
                        },
                        "evidenceReference": {
                          "type": "string"
                        }
                      },
                      "additionalProperties": false
                    }
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Soldes d'ouverture rapprochés et certifiés par Finance"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/accounting/runs/{runId}/post": {
      "post": {
        "operationId": "postApprovedCalculation",
        "parameters": [
          {
            "$ref": "#/components/parameters/CorrelationId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowApprovalCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Run comptabilisé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountingPostingResult"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/accounting/journals/{journalEntryId}/reverse": {
      "post": {
        "operationId": "reversePostedJournal",
        "parameters": [
          {
            "$ref": "#/components/parameters/CorrelationId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "name": "journalEntryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JournalReversalCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contrepassation comptabilisée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalReversalResult"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/accounting/events": {
      "post": {
        "operationId": "emitAccountingEvent",
        "parameters": [
          { "$ref": "#/components/parameters/CorrelationId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AccountingEventCommand" } } }
        },
        "responses": {
          "200": { "description": "Evenement comptable emis", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AccountingEventResult" } } } },
          "default": { "$ref": "#/components/responses/Problem" }
        }
      }
    },
    "/accounting/journals/{journalEntryId}/acknowledgements": {
      "post": {
        "operationId": "acknowledgeAccountingEvent",
        "parameters": [
          { "$ref": "#/components/parameters/CorrelationId" },
          { "$ref": "#/components/parameters/IdempotencyKey" },
          { "name": "journalEntryId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AccountingAcknowledgementCommand" } } }
        },
        "responses": {
          "200": { "description": "Accuse comptable mis a jour", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AccountingAcknowledgementResult" } } } },
          "default": { "$ref": "#/components/responses/Problem" }
        }
      }
    },
    "/accounting/reconciliations": {
      "post": {
        "operationId": "reconcileGeneralLedger",
        "parameters": [
          {
            "$ref": "#/components/parameters/CorrelationId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReconciliationCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rapprochement historisé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconciliationResult"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/compliance/sharia-reviews": {
      "post": {
        "operationId": "submitShariaReview",
        "parameters": [
          {
            "$ref": "#/components/parameters/CorrelationId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShariaReviewSubmission"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Dossier Charia soumis",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShariaReviewTransition"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/compliance/sharia-reviews/{reviewId}/review": {
      "post": {
        "operationId": "reviewShariaCase",
        "parameters": [
          {
            "$ref": "#/components/parameters/CorrelationId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "name": "reviewId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShariaOpinion"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Avis Charia enregistré",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShariaReviewTransition"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/compliance/sharia-reviews/{reviewId}/decide": {
      "post": {
        "operationId": "decideShariaCase",
        "parameters": [
          {
            "$ref": "#/components/parameters/CorrelationId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "name": "reviewId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShariaDecision"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Décision Charia enregistrée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShariaReviewTransition"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "keycloakBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "CorrelationId": {
        "name": "X-Correlation-ID",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 16,
          "maxLength": 128
        }
      }
    },
    "responses": {
      "Problem": {
        "description": "Erreur normalisée RFC 9457",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    },
    "schemas": {
...
      "WorkflowApprovalCommand": {
        "type": "object",
        "required": [
          "justification"
        ],
        "properties": {
          "justification": {
            "type": "string",
            "minLength": 10,
            "maxLength": 1000
          }
        },
        "additionalProperties": false
      },
...
      "AccountingPostingResult": {
        "type": "object",
        "required": [
          "state",
          "journalEntryId"
        ],
        "properties": {
          "state": {
            "type": "string",
            "const": "POSTED"
          },
          "journalEntryId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "JournalReversalCommand": {
        "type": "object",
        "required": [
          "justification",
          "reversalBusinessDate"
        ],
        "properties": {
          "justification": {
            "type": "string",
            "minLength": 10,
            "maxLength": 1000
          },
          "reversalBusinessDate": {
            "type": "string",
            "format": "date"
          }
        },
        "additionalProperties": false
      },
      "JournalReversalResult": {
        "type": "object",
        "required": [
          "state",
          "reversalJournalEntryId"
        ],
        "properties": {
          "state": {
            "type": "string",
            "const": "POSTED"
          },
          "reversalJournalEntryId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "AccountingJournalLine": {
        "type": "object",
        "required": ["accountCode", "currency", "debit", "credit"],
        "properties": {
          "accountCode": { "type": "string", "pattern": "^[A-Z0-9a-f:_-]{2,128}$" },
          "currency": { "type": "string", "pattern": "^[A-Z]{3}$" },
          "debit": { "type": "string", "pattern": "^[0-9]+(\\.[0-9]+)?$" },
          "credit": { "type": "string", "pattern": "^[0-9]+(\\.[0-9]+)?$" },
          "participantId": { "type": ["string", "null"], "minLength": 1, "maxLength": 128 }
        },
        "additionalProperties": false
      },
      "AccountingEventCommand": {
        "type": "object",
        "required": ["runId", "poolId", "productId", "eventType", "eventId", "entityId", "businessDate", "currencyScale", "lines"],
        "properties": {
          "runId": { "type": "string", "minLength": 1, "maxLength": 128 },
          "poolId": { "type": "string", "minLength": 1, "maxLength": 128 },
          "productId": { "type": "string", "minLength": 1, "maxLength": 128 },
          "eventType": { "type": "string", "enum": ["REVENUE", "PROFIT_SHARE", "PER_MOVEMENT", "IRR_MOVEMENT", "PURIFICATION", "WITHHOLDING_TAX", "REMAINDER", "CORRECTION"] },
          "eventId": { "type": "string", "minLength": 1, "maxLength": 128 },
          "entityId": { "type": "string", "minLength": 1, "maxLength": 128 },
          "businessDate": { "type": "string", "format": "date" },
          "currencyScale": { "type": "integer", "minimum": 0, "maximum": 6 },
          "lines": { "type": "array", "minItems": 2, "maxItems": 1000, "items": { "$ref": "#/components/schemas/AccountingJournalLine" } }
        },
        "additionalProperties": false
      },
      "AccountingEventResult": {
        "type": "object",
        "required": ["journalEntryId", "acknowledgementState"],
        "properties": {
          "journalEntryId": { "type": "string", "format": "uuid" },
          "acknowledgementState": { "type": "string", "const": "PENDING" }
        },
        "additionalProperties": false
      },
      "AccountingAcknowledgementCommand": {
        "type": "object",
        "required": ["action", "externalReference"],
        "properties": {
          "action": { "type": "string", "enum": ["ACKNOWLEDGED", "REJECTED", "RETRIED", "REVERSED"] },
          "externalReference": { "type": "string", "minLength": 1, "maxLength": 256 },
          "reason": { "type": "string", "minLength": 10, "maxLength": 2000 }
        },
        "additionalProperties": false
      },
      "AccountingAcknowledgementResult": {
        "type": "object",
        "required": ["acknowledgementState"],
        "properties": {
          "acknowledgementState": { "type": "string", "enum": ["PENDING", "ACKNOWLEDGED", "REJECTED", "RETRIED", "REVERSED"] }
        },
        "additionalProperties": false
      },
      "ReconciliationCommand": {
        "type": "object",
        "required": [
          "businessDate",
          "currency",
          "generalLedgerAmount",
          "sourceReference",
          "sourceChecksumSha256"
        ],
        "properties": {
          "businessDate": {
            "type": "string",
            "format": "date"
          },
          "currency": {
            "type": "string",
            "pattern": "^[A-Z]{3}$"
          },
          "generalLedgerAmount": {
            "type": "string",
            "pattern": "^-?[0-9]+(\\.[0-9]+)?$"
          },
          "sourceReference": {
            "type": "string",
            "minLength": 3,
            "maxLength": 256
          },
          "sourceChecksumSha256": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$"
          }
        },
        "additionalProperties": false
      },
      "ReconciliationResult": {
        "type": "object",
        "required": [
          "reconciliationId",
          "state",
          "difference"
        ],
        "properties": {
          "reconciliationId": {
            "type": "string",
            "format": "uuid"
          },
          "state": {
            "type": "string",
            "enum": [
              "MATCHED",
              "VARIANCE"
            ]
          },
          "difference": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ShariaReviewSubmission": {
        "type": "object",
        "required": [
          "resourceType",
          "resourceId"
        ],
        "properties": {
          "resourceType": {
            "type": "string",
            "minLength": 2,
            "maxLength": 64
          },
          "resourceId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          }
        },
        "additionalProperties": false
      },
      "ShariaOpinion": {
        "type": "object",
        "required": [
          "opinion"
        ],
        "properties": {
          "opinion": {
            "type": "string",
            "minLength": 10,
            "maxLength": 4000
          }
        },
        "additionalProperties": false
      },
      "ShariaDecision": {
        "type": "object",
        "required": [
          "decision",
          "justification",
          "evidenceDocumentId"
        ],
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "APPROVED",
              "REJECTED"
            ]
          },
          "justification": {
            "type": "string",
            "minLength": 10,
            "maxLength": 4000
          },
          "evidenceDocumentId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "ShariaReviewTransition": {
        "type": "object",
        "required": [
          "reviewId",
          "state"
        ],
        "properties": {
          "reviewId": {
            "type": "string",
            "format": "uuid"
          },
          "state": {
            "type": "string",
            "enum": [
              "SUBMITTED",
              "REVIEWED",
              "APPROVED",
              "REJECTED"
            ]
          }
        },
        "additionalProperties": false
      },
...
      "Problem": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "format": "uri-reference"
          },
          "correlationId": {
            "type": "string",
            "format": "uuid"
          }
        }
      }
    }
  }
}